+++ Server_Root/Server/RoomServer/Behavior/Behaviors/Dash.h
@@ -20,7 +20,7 @@ public:
 	static const string NAME;
 
 public:
-	void Start(__int64 lStartTime, ...) override;
+	void Start(long long lStartTime, ...) override;
 	void Initialize() override;
-	void UpdateBody(__int64 lUpdateTime) override;
+	void UpdateBody(long long lUpdateTime) override;
 };
