+++ Server_Root/Server/RoomServer/Skill/ISkill.h
@@ -24,9 +24,9 @@ protected:
 	bool	m_bActivated = false;
 
 protected:
-	__int64	m_lStartTime;			//	ElapsedTime after game was started (Milliseconds)
-	__int64	m_lLastUpdateTime;		//	ElapsedTime after game was started (Milliseconds)
-	__int64	m_lEndTime = -1;		//	ElapsedTime after game was started (Milliseconds)
+	long long m_lStartTime;			//	ElapsedTime after game was started (Milliseconds)
+	long long m_lLastUpdateTime;	//	ElapsedTime after game was started (Milliseconds)
+	long long m_lEndTime = -1;		//	ElapsedTime after game was started (Milliseconds)
 	float m_fDeltaTime;				//	CurrentTime - PreviousTime (seconds)
 	float m_fCurrentTime;			//	ElapsedTime after behavior was started (seconds)
 	float m_fPreviousTime;			//	ElapsedTime after behavior was started (seconds)
@@ -35,18 +35,18 @@ public:
 	int GetMasterDataID();
 
 public:
-	void Update(__int64 lUpdateTime);
-	void Stop(__int64 lTime);
+	void Update(long long lUpdateTime);
+	void Stop(long long lTime);
 
 protected:
-	virtual bool IsCoolTime(__int64 lTime) = 0;
-	virtual bool IsValidToStart(__int64 lTime) = 0;
-	virtual void Start(__int64 lStartTime, ...);
+	virtual bool IsCoolTime(long long lTime) = 0;
+	virtual bool IsValidToStart(long long lTime) = 0;
+	virtual void Start(long long lStartTime, ...);
 
 public:
 	virtual void Initialize() = 0;
-	virtual void UpdateBody(__int64 lUpdateTime) = 0;
-	virtual void ProcessInput(__int64 lTime, BaeGameRoom* pBaeGameRoom, GameInputSkillToR* pMsg) = 0;
+	virtual void UpdateBody(long long lUpdateTime) = 0;
+	virtual void ProcessInput(long long lTime, BaeGameRoom* pBaeGameRoom, GameInputSkillToR* pMsg) = 0;
 
 public:
 	float GetTime();
