+++ Server_Root/Server/RoomServer/Skill/Skills/FireSkill.h
@@ -26,11 +26,11 @@ private:
 	float	m_fMP;
 
 protected:
-	bool IsCoolTime(__int64 lTime) override;
-	bool IsValidToStart(__int64 lTime) override;
+	bool IsCoolTime(long long lTime) override;
+	bool IsValidToStart(long long lTime) override;
 
 public:
 	void Initialize() override;
-	void UpdateBody(__int64 lUpdateTime) override;
-	void ProcessInput(__int64 lTime, BaeGameRoom* pBaeGameRoom, GameInputSkillToR* pMsg) override;
+	void UpdateBody(long long lUpdateTime) override;
+	void ProcessInput(long long lTime, BaeGameRoom* pBaeGameRoom, GameInputSkillToR* pMsg) override;
 };
