+++ Server_Root/Server/RoomServer/State/States/Shield.h
@@ -9,7 +9,7 @@ using namespace std;
 class Shield : public IState
 {
 public:
-	Shield(IEntity* pEntity, int nMasterDataID, __int64 lStartTime);
+	Shield(IEntity* pEntity, int nMasterDataID, long long lStartTime);
 	virtual ~Shield();
 
 public:
@@ -24,5 +24,5 @@ private:
 public:
 	int GetID() override;
 	void Initialize() override;
-	void UpdateBody(__int64 lUpdateTime) override;
+	void UpdateBody(long long lUpdateTime) override;
 };
